[fix] notification field user id change type on uint

+++ internal/services/notifier/notification.go
@@ -1,7 +1,7 @@
 package notifier
 
 type Notification struct {
-	FromUserID uint64
+	FromUserID uint
 	Key        string
 	Payload    interface{}
 }
